change a little bit

+++ finalreducer.py
@@ -9,22 +9,20 @@ tweet = None
 
 # input comes from STDIN
 for line in sys.stdin:
-    # remove leading and trailing whitespace
+
     line = line.strip()
 
-    # parse the input we got from mapper.py
+
     tweet, rank = line.split(',')
 
-    # convert count (currently a string) to int
+
     try:
-        rank = int(rank)
+        rank = float(rank)
     except ValueError:
-        # count was not a number, so silently
-        # ignore/discard this line
+
         continue
 
-    # this IF-switch only works because Hadoop sorts map output
-    # by key (here: word) before it is passed to the reducer
+
     if current_tweet == tweet:
         current_rank += rank
     else:
